+++ All.ui/miniprogram/components/timePicker/index.js
@@ -1,0 +1,26 @@
+// components/timePicker/index.js
+Component({
+  /**
+   * 组件的属性列表
+   */
+  properties: {
+
+  },
+
+  /**
+   * 组件的初始数据
+   */
+  data: {
+
+  },
+
+  /**
+   * 组件的方法列表
+   */
+  methods: {
+
+    open: function() {
+      this.selectComponent('#drawerMoto').open()
+    }
+  }
+})